상품 선택화면 수정, 상품 수정 삭제기능 추가

+++ src/main/java/com/multi/mini/product/controller/ProductController.java
@@ -16,7 +16,6 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
 
@@ -53,7 +52,7 @@ public class ProductController {
 
 
     @PostMapping("insert")
-    public void productInsert(ProductDTO productDTO, @RequestParam(value = "file") MultipartFile file, HttpServletRequest request) throws IOException {
+    public void productInsert(ProductDTO productDTO, @RequestParam(value = "file") MultipartFile file, HttpServletRequest request) throws Exception {
 
         String path = request.getSession().getServletContext().getRealPath("");
         String root = path.substring(0,path.length() - 7);
@@ -67,24 +66,95 @@ public class ProductController {
             mkdir.mkdirs();
         }
 
-        String originFileName = file.getOriginalFilename();
-        String ext = originFileName.substring(originFileName.lastIndexOf("."));
-        String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
+        if (!file.isEmpty()){
+            String originFileName = file.getOriginalFilename();
+            String ext = originFileName.substring(originFileName.lastIndexOf("."));
+            String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
 
-        try {
-            file.transferTo(new File(filePath + "\\" + savedName));
+            try {
+                file.transferTo(new File(filePath + "\\" + savedName));
 
-            productDTO.setProductImg(savedName);
+                productDTO.setProductImg(savedName);
 
+                int result = productService.insertProduct(productDTO);
+                System.out.println("인서트 결과" + result);
+
+            }catch (Exception e) {
+                e.printStackTrace();
+                /* 실패시 파일 삭제 */
+                new File(filePath + "\\" + savedName).delete();
+            }
+        }
+        else {
+            System.out.println("이미지 없이 생성");
             int result = productService.insertProduct(productDTO);
             System.out.println(result);
+        }
 
-        }catch (Exception e) {
-            e.printStackTrace();
-            /* 실패시 파일 삭제 */
-            new File(filePath + "\\" + savedName).delete();
+    }
+
+    @PostMapping("update")
+    public void productUpdate(ProductDTO productDTO, @RequestParam(value = "file") MultipartFile file, HttpServletRequest request) throws Exception{
+
+        String path = request.getSession().getServletContext().getRealPath("");
+        String root = path.substring(0,path.length() - 7);
+        System.out.println("path : " + path);
+        System.out.println("root : " + root);
+
+        String filePath = root + "resources/static/img/uploadFiles";
+
+
+
+        File mkdir = new File(filePath);
+        if (!mkdir.exists()) {
+            mkdir.mkdirs();
+        }
+
+        if (!file.isEmpty()){
+            String originFileName = file.getOriginalFilename();
+            String ext = originFileName.substring(originFileName.lastIndexOf("."));
+            String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
+
+            try {
+                //기존 이미지 삭제
+                new File(filePath + "\\" + productDTO.getProductImg()).delete();
+
+                file.transferTo(new File(filePath + "\\" + savedName));
+
+                productDTO.setProductImg(savedName);
+
+                int result = productService.updateProduct(productDTO);
+                System.out.println("업데이트 결과" + result);
+
+            }catch (Exception e) {
+                e.printStackTrace();
+                /* 실패시 파일 삭제 */
+                new File(filePath + "\\" + savedName).delete();
+            }
+        }
+        else {
+            System.out.println("이미지 없이 변경");
+            int result = productService.updateProduct(productDTO);
+            System.out.println("업데이트 결과" + result);
         }
     }
+
+    @PostMapping("delete")
+    public void productDelete (ProductDTO productDTO , HttpServletRequest request) throws Exception {
+
+        String path = request.getSession().getServletContext().getRealPath("");
+        String root = path.substring(0,path.length() - 7);
+        String filePath = root + "resources/static/img/uploadFiles";
+
+        //기존 이미지 삭제
+        new File(filePath + "\\" + productDTO.getProductImg()).delete();
+
+
+        int result = productService.deleteProduct(productDTO.getProductNo());
+        System.out.println("딜리트 결과" + result);
+    }
+
+
 
     @GetMapping(value="productbycategory", produces = "application/json; charset=UTF-8")
     @ResponseBody
@@ -121,6 +191,21 @@ public class ProductController {
         session.setAttribute( productListDTO.getMemberNo() + "basket" , productListDTO);
 
 
+    }
+
+    @GetMapping(value="getbasket", produces = "application/json; charset=UTF-8")
+    @ResponseBody
+    public ProductListDTO getbasket(HttpSession session){
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+
+        ProductListDTO basket = (ProductListDTO)session.getAttribute(userDetails.getMemberNo()+"basket");
+
+        System.out.println("basket>> " + basket);
+
+        return basket;
     }
 
 }
